Projects/AerTrip-Solution/requirements.txt
absl-py==0.9.0
aiofiles==0.4.0
aiohttp==3.6.2
alabaster==0.7.12
amqp==2.5.2
appdirs==1.4.3
appnope==0.1.0
APScheduler==3.6.3
astor==0.8.1
astroid==2.4.2
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.1.0
Automat==20.2.0
autopep8==1.5
babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.0
billiard==3.6.3.0
black==19.10b0
bleach==3.1.0
blis==0.4.1
boto==2.49.0
boto3==1.11.8
botocore==1.14.8
bpemb==0.3.0
bz2file==0.98
cachetools==4.0.0
catalogue==1.0.0
celery==4.4.1
certifi==2019.11.28
cffi==1.13.1
chardet==3.0.4
chatette==1.6.2
clean-text==0.2.1
Click==7.0
cloudpickle==1.2.2
colorama==0.4.3
colorclass==2.2.0
coloredlogs==10.0
colorhash==1.0.2
commonmark==0.9.1
ConfigArgParse==1.0
constantly==15.1.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
Cython==0.29.14
cytoolz==0.9.0.1
dataclasses==0.7
datefinder==0.7.1
decorator==4.4.0
defusedxml==0.6.0
Deprecated==1.2.9
dill==0.2.9
distlib==0.3.0
dnspython==1.15.0
docformatter==1.3.1
docopt==0.6.2
docutils==0.15.2
dopamine-rl==3.0.1
emoji==0.5.4
en-core-web-md==2.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
enum-compat==0.0.3
Faker==0.8.13
fbmessenger==6.0.0
filelock==3.0.12
flair @ git+https://github.com/flairNLP/flair.git@f5d846ffcaf56d9f90cde484b234685c17a23096
flake8==3.7.8
flake8-bugbear==20.1.4
Flask==0.12.4
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-JWT-Extended==3.7.0
Flask-Login==0.4.0
ftfy==5.8
future==0.17.1
gast==0.2.2
gensim==3.8.3
gevent==1.4.0
gin-config==0.3.0
gitdb==4.0.5
gitdb2==4.0.2
gitpython==3.1.3
google-api-python-client==1.7.11
google-auth==1.11.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.26.0
gunicorn==20.0.4
gym==0.15.4
h11==0.8.1
h2==3.1.1
h5py==2.10.0
hpack==3.0.0
html2text==2020.1.16
httpcore==0.3.0
httplib2==0.16.0
httptools==0.0.13
humanfriendly==4.18
hyperframe==5.2.0
hyperlink==19.0.0
hyperopt==0.2.4
idna==2.8
idna-ssl==1.1.0
imagesize==1.2.0
importlib-metadata==0.23
importlib-resources==1.0.2
incremental==17.5.0
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jovian==0.2.14
jsonpickle==1.2
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kfac==0.2.0
kiwisolver==1.1.0
klein==17.10.0
kombu==4.6.8
langdetect==1.0.8
lazy-object-proxy==1.4.3
line-bot-sdk==1.16.0
line-bot-sdk-extra==0.1.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mattermostwrapper==2.1
mccabe==0.6.1
mesh-tensorflow==0.1.9
mistune==0.8.4
mock==4.0.1
more-itertools==7.2.0
mpld3==0.3
mpmath==1.1.0
msgpack==0.5.6
msgpack-numpy==0.4.3.2
multidict==4.6.1
murmurhash==1.0.2
mypy-extensions==0.4.3
mysql==0.0.2
mysqlclient==2.0.1
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.4
nltk==3.4.5
notebook==6.0.1
numpy==1.18.1
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.1.2.30
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.7.0
pexpect==4.7.0
PGen==0.2.1
pickleshare==0.7.5
pika==1.0.1
Pillow==7.0.0
plac==0.9.6
pluggy==0.6.0
preshed==3.0.4
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==2.0.10
protobuf==3.11.2
ptyprocess==0.6.0
py==1.8.0
py-gfm==0.1.4
py-rouge==1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
pydot==1.4.1
pyflakes==2.1.1
pyglet==1.3.2
Pygments==2.4.2
PyHamcrest==2.0.1
PyJWT==1.7.1
pykwalify==1.7.0
pylint==2.5.3
pymongo==3.7.2
PyMySQL==0.10.0
pyparsing==2.4.6
pypng==0.0.20
pyreadline==2.1
pyrsistent==0.15.4
PySocks==1.7.1
pyspellchecker==0.5.4
pytest==3.3.0
pytest-flask==0.11.0
python-crfsuite==0.9.6
python-dateutil==2.8.0
python-dotenv==0.12.0
python-engineio==3.11.2
python-socketio==4.4.0
python-telegram-bot==11.1.0
pytz==2019.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3
pyzmq==18.1.0
qtconsole==4.5.5
questionary==1.5.1
rabbitmq==0.2.0
rasa==1.7.4
rasa-nlu==0.15.1
rasa-sdk==1.7.0
recommonmark==0.6.0
redis==3.3.11
regex==2018.1.10
requests==2.22.0
requests-async==0.5.0
requests-mock==1.8.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.3.2
rocketchat-API==0.6.36
rsa==4.0
ruamel.yaml==0.15.100
s3transfer==0.3.2
sacremoses==0.0.43
sanic==19.9.0
Sanic-Cors==0.9.9.post1
sanic-jwt==1.4.0
Sanic-Plugins-Framework==0.8.2.post1
scikit-learn==0.22.2.post1
scipy==1.4.1
segtok==1.5.9
Send2Trash==1.5.0
sentencepiece==0.1.86
sh==1.13.1
simplejson==3.17.0
six==1.12.0
sklearn-crfsuite==0.3.6
slackclient==1.3.2
smart-open==2.0.0
snowballstemmer==2.0.0
soupsieve==2.0
spacy==2.3.2
spello==1.2.0
sphinx==3.1.2
sphinx-autodoc-typehints==1.11.0
sphinx-rtd-theme==0.5.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.13
sqlitedict==1.6.0
srsly==1.0.4
subword-nmt @ https://github.com/rsennrich/subword-nmt/archive/master.zip
sympy==1.5.1
tabulate==0.8.6
tensor2tensor==1.14.1
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-cpu==1.15.0
tensorflow-cpu-estimator==1.15.1
tensorflow-datasets==1.3.2
tensorflow-estimator==1.15.1
tensorflow-gan==2.0.0
tensorflow-hub==0.7.0
tensorflow-metadata==0.21.0
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.2
terminaltables==3.1.0
testpath==0.4.2
text-unidecode==1.2
textblob==0.15.3
thinc==7.4.1
tokenizers==0.5.2
toml==0.10.0
toolz==0.10.0
torch==1.5.0+cpu
torchvision==0.6.0+cpu
tornado==6.0.3
tqdm==4.41.1
traitlets==4.3.3
transformers==2.8.0
tweepy==3.9.0
twilio==6.35.3
Twisted==19.10.0
typed-ast==1.4.1
typing==3.7.4.1
typing-extensions==3.7.4.1
tzlocal==2.0.0
ujson==1.35
unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.24.3
uuid==1.30
vine==1.3.0
virtualenv==20.0.7
wasabi==0.8.0
wcwidth==0.1.7
webencodings==0.5.1
webexteamssdk==1.2
websocket-client==0.54.0
websocket-server==0.4
websockets==8.0.2
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.12.0
xlrd==1.2.0
yarl==1.4.2
zipp==0.6.0
zope.interface==4.7.1
